+++ src/fields.rs
@@ -1,54 +1,52 @@
-
 use argh::FromArgValue;
 
 /// one of the tables that can be displayed
-#[derive(Debug, Clone, Copy)]
-pub enum Table {
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum Field {
     Dates,
     Status,
     RemoteAddresses,
     Referers,
     Paths, // popular paths
     Methods,
-    Trends, // trendy paths
 }
 
-pub static DEFAULT_TABLES: &[Table] = &[
-    Table::Dates,
-    Table::Status,
-    Table::RemoteAddresses,
-    Table::Referers,
-    Table::Paths,
-    Table::Trends,
+pub static DEFAULT_TABLES: &[Field] = &[
+    Field::Dates,
+    Field::Status,
+    Field::Referers,
+    Field::Paths,
 ];
 
-pub static ALL_TABLES: &[Table] = &[
-    Table::Dates,
-    Table::Methods,
-    Table::Status,
-    Table::RemoteAddresses,
-    Table::Referers,
-    Table::Paths,
-    Table::Trends,
+pub static ALL_TABLES: &[Field] = &[
+    Field::Dates,
+    Field::Methods,
+    Field::Status,
+    Field::RemoteAddresses,
+    Field::Referers,
+    Field::Paths,
 ];
 
 #[derive(Debug, Clone)]
-pub struct Tables(Vec<Table>);
+pub struct Fields(Vec<Field>);
 
-impl Default for Tables {
+impl Default for Fields {
     fn default() -> Self {
         Self (DEFAULT_TABLES.to_vec())
     }
 }
 
-impl Tables {
+impl Fields {
     fn all() -> Self {
         Self (ALL_TABLES.to_vec())
     }
+    pub fn contains(&self, tbl: Field) -> bool {
+        self.0.contains(&tbl)
+    }
 }
 
-impl IntoIterator for Tables {
-    type Item = Table;
+impl IntoIterator for Fields {
+    type Item = Field;
     type IntoIter = std::vec::IntoIter<Self::Item>;
 
     fn into_iter(self) -> Self::IntoIter {
@@ -56,7 +54,7 @@ impl IntoIterator for Tables {
     }
 }
 
-impl FromArgValue for Tables {
+impl FromArgValue for Fields {
     fn from_arg_value(value: &str) -> Result<Self, String> {
         let value = value.to_lowercase();
         if value == "all" || value == "a" {
@@ -65,19 +63,17 @@ impl FromArgValue for Tables {
         let mut v = Vec::new();
         for s in value.split(',') {
             if s.contains("date") {
-                v.push(Table::Dates);
+                v.push(Field::Dates);
             } else if s.contains("stat") {
-                v.push(Table::Status);
-            } else if s.contains("addr") {
-                v.push(Table::RemoteAddresses);
+                v.push(Field::Status);
+            } else if s.contains("addr") || s.contains("ip") {
+                v.push(Field::RemoteAddresses);
             } else if s.contains("ref") {
-                v.push(Table::Referers);
+                v.push(Field::Referers);
             } else if s.contains("path") {
-                v.push(Table::Paths);
+                v.push(Field::Paths);
             } else if s.contains("method") {
-                v.push(Table::Methods);
-            } else if s.contains("trend") {
-                v.push(Table::Trends);
+                v.push(Field::Methods);
             } else {
                 return Err(format!("Unrecognized table : {:?}", s));
             }