+++ src/line_group.rs
@@ -1,18 +1,30 @@
 use {
     crate::*,
+    std::{
+        fmt::Display,
+        hash::Hash,
+    },
 };
 
 /// A non empty group of lines, with a common characteristic,
 /// for stats
-pub struct LineGroup<'b> {
+pub struct LineGroup<'b, T>
+where
+        T: Display + Hash + Eq + 'b,
+{
+    pub value: T,
     pub lines: Vec<&'b LogLine>, // guaranteed not empty
     pub trend: Trend,
     pub bytes: u64,
     pub key_sum: u64,
 }
 
-impl<'b> LineGroup<'b> {
+impl<'b, T> LineGroup<'b, T>
+where
+        T: Display + Hash + Eq + 'b,
+{
     pub fn new(
+        value: T,
         lines: Vec<&'b LogLine>,
         trend_computer: &TrendComputer,
     ) -> Self {
@@ -24,6 +36,7 @@ impl<'b> LineGroup<'b> {
             Key::Bytes => bytes,
         };
         Self {
+            value,
             lines,
             trend,
             bytes,