+++ src/md/addr.rs
@@ -8,20 +8,23 @@ pub fn print_remote_addresses(
     printer: &Printer,
     trend_computer: Option<&TrendComputer>,
 ) {
-    let n = match printer.detail_level {
+    let limit = match printer.detail_level {
         0 => 3,
         1 => 5,
         l => l * 10,
     };
+    let section = Section {
+        groups_name: "remote IP addresses",
+        group_key: "IP address",
+        view: View::Limited(limit),
+        changes: true,
+    };
     printer.print_groups(
-        "remote addresses",
-        "remote address",
+        &section,
         log_lines,
         |_| true,
         |line| line.remote_addr,
         trend_computer,
-        n,
-        true,
     );
 }