+++ src/md/mod.rs
@@ -5,12 +5,13 @@ pub mod summary;
 mod status;
 mod methods;
 mod printer;
+mod section;
 mod skin;
 
 pub use {
     printer::*,
+    section::*,
 };
-
 
 use {
     crate::*,
@@ -25,48 +26,22 @@ pub fn print_analysis(
         return;
     }
     let log_lines = &log_base.lines;
-    let mut popular_paths = false;
-    let mut trendy_paths = false;
-    for table in printer.tables.clone().into_iter() {
-        match table {
-            Table::Dates => {
-                let histogram = Histogram::from(&log_base);
-                histogram.print(printer);
-            }
-            Table::Status => {
-                status::print_status_codes(log_lines, printer, trend_computer);
-            }
-            Table::RemoteAddresses => {
-                addr::print_remote_addresses(log_lines, printer, trend_computer);
-            }
-            Table::Referers => {
-                referers::print_referers(log_lines, printer, trend_computer);
-            }
-            Table::Paths => {
-                popular_paths = true;
-            }
-            Table::Trends => {
-                trendy_paths = true;
-            }
-            Table::Methods => {
-                methods::print_methods(log_lines, printer, trend_computer);
-            }
-        }
+    if printer.fields.contains(Field::Dates) {
+        Histogram::from(&log_base).print(printer);
     }
-    if popular_paths || trendy_paths {
-        if let Some(trend_computer) = trend_computer {
-            paths::print_paths(
-                &log_base,
-                printer,
-                trend_computer,
-                popular_paths,
-                trendy_paths,
-            );
-        } else {
-            paths::print_paths_no_trends(
-                &log_base,
-                printer,
-            );
-        }
+    if printer.fields.contains(Field::Methods) {
+        methods::print_methods(log_lines, printer, trend_computer);
+    }
+    if printer.fields.contains(Field::Status) {
+        status::print_status_codes(log_lines, printer, trend_computer);
+    }
+    if printer.fields.contains(Field::RemoteAddresses) {
+        addr::print_remote_addresses(log_lines, printer, trend_computer);
+    }
+    if printer.fields.contains(Field::Referers) {
+        referers::print_referers(log_lines, printer, trend_computer);
+    }
+    if printer.fields.contains(Field::Paths) {
+        paths::print_paths(log_lines, printer, trend_computer);
     }
 }